fixed the error of 3.14 and float and explained 2 ways to do it.

+++ src/HelloWorld.java
@@ -4,7 +4,7 @@ public class HelloWorld {
         System.out.println("World!");
         int myFavoriteNumber = 21;System.out.println(myFavoriteNumber);
         String myString = "My Favorite Number";System.out.println(myString);//doesn't work with char because it's not compatible.
-        float myNumber = 3.14;
+        float myNumber = (float) 3.14; // 2 ways to fix this is 1. double instead of float as the data type or 2. (float) before the value.
         System.out.print(myNumber);
     }
 }
